Fixed Program::step(), improved setting parse mode
`Program::step()` was set to return `Program::is_done()`, which is
backwards from what it should've been. Setting it to return
`!Program::is_done()` allows one to write code like:
```
while program.step()
{
   // Some stuff to do between steps
}
```

Improved setting the parse mode. Originally, only basic and Extended
Type I was going to be supported, and `parse_code()` was written with
that in mind. An enum of different modes has been created, and now
`parse_code()` takes that enum instead of a bool for setting the mode to
parse in.