Fixed issue with piped input
Before, if input to a program was piped in and the program read input
until a nul character was read, the program would crash due to an
unexpected end of input.
Now if the program tries to read input and there is no input left, the
byte that is to be set will be set to 0, allowing code such as `,[.,]`
to work as expected.