Began rework to support self-modifying languages
Due to the way that brfk was originally written, variants of Branfuck
that were self-modifying would not work. This is a rework of brfk that
will A) be able to support those languages and B) add an API that anyone
can use without having to use the brfk executable.

This rework is not yet finished or functional.